third_party/WebKit/LayoutTests/http/tests/inspector/compiler-script-mapping-expected.txt

Tests SourceMap and CompilerScriptMapping.


Running: testSimpleMap
example.js === example.js
0 === 0
9 === 9
example.js === example.js
0 === 0
13 === 13
example.js === example.js
0 === 0
25 === 25
example.js === example.js
2 === 2
4 === 4
example.js === example.js
2 === 2
11 === 11
example.js === example.js
2 === 2
24 === 24
undefined === undefined
undefined === undefined
undefined === undefined
0 === 0
0 === 0
0 === 0
17 === 17
0 === 0
18 === 18
source line 4 has no mappings.
0 === 0
29 === 29

Running: testNoMappingEntry
example.js === example.js
0 === 0
0 === 0
example.js === example.js
0 === 0
2 === 2

Running: testEmptyLine
example.js === example.js
0 === 0
0 === 0
3 === 3
1 === 1

Running: testSections
source1.js === source1.js
0 === 0
0 === 0
source1.js === source1.js
2 === 2
1 === 1
source2.js === source2.js
0 === 0
0 === 0
source2.js === source2.js
2 === 2
1 === 1

Running: testResolveSourceMapURL
http://example.com/map.json === http://example.com/map.json
http://example.com/map.json === http://example.com/map.json
http://example.com/maps/map.json === http://example.com/maps/map.json
Tests from http://tools.ietf.org/html/rfc3986#section-5.4 using baseURL="http://a/b/c/d;p?q"
http://h resolves to http://h===http://h passes: true
g resolves to http://a/b/c/g===http://a/b/c/g passes: true
./g resolves to http://a/b/c/g===http://a/b/c/g passes: true
g/ resolves to http://a/b/c/g/===http://a/b/c/g/ passes: true
/g resolves to http://a/g===http://a/g passes: true
//g resolves to http://g===http://g passes: true
?y resolves to http://a/b/c/d;p?y===http://a/b/c/d;p?y passes: true
g?y resolves to http://a/b/c/g?y===http://a/b/c/g?y passes: true
#s resolves to http://a/b/c/d;p?q#s===http://a/b/c/d;p?q#s passes: true
g#s resolves to http://a/b/c/g#s===http://a/b/c/g#s passes: true
g?y#s resolves to http://a/b/c/g?y#s===http://a/b/c/g?y#s passes: true
;x resolves to http://a/b/c/;x===http://a/b/c/;x passes: true
g;x resolves to http://a/b/c/g;x===http://a/b/c/g;x passes: true
g;x?y#s resolves to http://a/b/c/g;x?y#s===http://a/b/c/g;x?y#s passes: true
 resolves to http://a/b/c/d;p?q===http://a/b/c/d;p?q passes: true
. resolves to http://a/b/c/===http://a/b/c/ passes: true
./ resolves to http://a/b/c/===http://a/b/c/ passes: true
.. resolves to http://a/b/===http://a/b/ passes: true
../ resolves to http://a/b/===http://a/b/ passes: true
../g resolves to http://a/b/g===http://a/b/g passes: true
../.. resolves to http://a/===http://a/ passes: true
../../ resolves to http://a/===http://a/ passes: true
../../g resolves to http://a/g===http://a/g passes: true
../../../g resolves to http://a/g===http://a/g passes: true
../../../../g resolves to http://a/g===http://a/g passes: true
/./g resolves to http://a/g===http://a/g passes: true
/../g resolves to http://a/g===http://a/g passes: true
g. resolves to http://a/b/c/g.===http://a/b/c/g. passes: true
.g resolves to http://a/b/c/.g===http://a/b/c/.g passes: true
g.. resolves to http://a/b/c/g..===http://a/b/c/g.. passes: true
..g resolves to http://a/b/c/..g===http://a/b/c/..g passes: true
./../g resolves to http://a/b/g===http://a/b/g passes: true
./g/. resolves to http://a/b/c/g/===http://a/b/c/g/ passes: true
g/./h resolves to http://a/b/c/g/h===http://a/b/c/g/h passes: true
g/../h resolves to http://a/b/c/h===http://a/b/c/h passes: true
g;x=1/./y resolves to http://a/b/c/g;x=1/y===http://a/b/c/g;x=1/y passes: true
g;x=1/../y resolves to http://a/b/c/y===http://a/b/c/y passes: true
g?y/./x resolves to http://a/b/c/g?y/./x===http://a/b/c/g?y/./x passes: true
g?y/../x resolves to http://a/b/c/g?y/../x===http://a/b/c/g?y/../x passes: true
g#s/./x resolves to http://a/b/c/g#s/./x===http://a/b/c/g#s/./x passes: true
g#s/../x resolves to http://a/b/c/g#s/../x===http://a/b/c/g#s/../x passes: true
Custom completeURL tests
//secure.com/moo resolves to http://secure.com/moo===http://secure.com/moo passes: true
cat.jpeg resolves to http://a/b/c/cat.jpeg===http://a/b/c/cat.jpeg passes: true
 resolves to http://example.com/path.css?query===http://example.com/path.css?query passes: true

Running: testCompilerScriptMapping
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source1.js UISourceCode arrived
source2.js UISourceCode arrived
Location checks passed. Requesting content
Content1 arrived.
Content2 arrived.

Running: testInlinedSourceMap
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source3.js UISourceCode arrived
Location checks passed. Requesting content
<source content> === <source content>

Running: testSourceMapCouldNotBeLoaded
Adding compiled.js
compiled.js UISourceCode arrived
compiled.js script found
source-map.json_ UISourceCode arrived
debugger:///VM

Running: testSourceRoot
/example.js === /example.js
0 === 0
9 === 9
0 === 0
0 === 0

Running: testNameClash
example.js === example.js [sm]
0 === 0
9 === 9
source line 0 has no mappings.

Running: testNameIndexes
1:0 > [no name assigned]
3:0 > [no name assigned]
3:18 > [no name assigned]
3:4 > 'name1'
4:0 > [no name assigned]
4:0 > [no name assigned]
4:4 > 'generated31465'
4:27 > [no name assigned]
5:0 > [no name assigned]
5:0 > [no name assigned]
